+++ R/define.modules.r
@@ -14,23 +14,19 @@
 #' landmarks for module 2. Repeat until all modules are defined.
 #' }
 #' 
-#'  \subsection{Selection in 3D}{ 
-#' Choosing which landmarks will be included in each module involves landmark selection using a mouse in 
-#' the rgl plot window. The user is prompted to select one or more landmarks. To do so, use the RIGHT mouse button 
-#' (or command + LEFT button for Mac users), draw a rectangle around landmarks to select.
-#' Selected landmarks will be colored yellow. Then type into the console a letter (e.g. 1, 2, 3...) to assign selected landmark(s) 
-#' to this module. Repeat until all landmarks are assigned to modules.
-#' }
+#'  \subsection{Notes for geomorph 4.1}{ 
+#'  Starting with geomorph version 4.1, interactive module selection in 3D is no longer
+#'  supported, as RGL is not supported under MacOS Tahoe.
+#'  }
 #' 
-#' @param spec A p x k matrix containing landmark coordinates of a single specimen (2D or 3D)
+#' @param spec A p x k matrix containing landmark coordinates of a single specimen (2D)
 #' @param nmodules Number of modules to be defined
 #' @return Function returns a vector of which landmarks belong in which module (e.g. 1, 1, 1, 2, 2, 3, 3, 3, 2) to be used
 #' with \code{\link{modularity.test}} or \code{\link{integration.test}}.
 #' @export
 #' @keywords utilities
 #' @seealso  \code{\link{modularity.test}} and \code{\link{integration.test}} 
-#' @author Emma Sherratt
-#' @seealso  \code{\link[rgl]{rgl-package}} (used in 3D plotting)
+#' @author Dean Adams and Emma Sherratt
 #' 
 define.modules <- function(spec, nmodules){
   spec.name <- deparse(substitute(spec))
@@ -55,40 +51,8 @@ define.modules <- function(spec, nmodules){
       select <- identifyPch(spec, col=module[i,2])
       selected[select] <- module[i,1] }
   }
-  # 3D
   if (checkdim == 3) {
-    plot3d(spec[, 1], spec[, 2],spec[, 3], size = 5,
-         xlim = range(spec[, 1]), ylim = range(spec[, 2]), zlim = range(spec[, 3]), 
-         asp = 1, box=F, axes=F, xlab="", ylab="", zlab="")
-    text3d(spec[, 1], spec[, 2], spec[, 3], texts = paste(1:dim(spec)[1]), 
-         adj = 1.3, pos = 4)
-    rgl.bringtotop(stay = FALSE)
-    while(anyNA(selected)==TRUE){
-      cat("Select landmarks","\n")
-      f<-select<-ans<-NULL
-      f<-select3d(button="right")
-      select<-f(spec)
-      if(anyNA(spec[which(select==TRUE)[1],])==TRUE){
-            cat("No vertex selected, try again","\n") 
-            f<-select<-NULL
-            f<-select3d(button="right")
-            select<-f(spec)}
-      points3d(spec[which(select==TRUE),1],spec[which(select==TRUE),2],spec[which(select==TRUE),3],
-               size=12,color="yellow")
-      cat(paste("Assign landmarks", paste(which(select==TRUE), collapse=","), "to which module?","\n")) 
-      ans<-readLines(n=1)
-      if(!is.null(ans)){
-        selected[select]<-ans
-        plot3d(spec[, 1], spec[, 2],spec[, 3], size = 5,
-               xlim = range(spec[, 1]), ylim = range(spec[, 2]), zlim = range(spec[, 3]), 
-               asp = 1, box=F, axes=F, xlab="", ylab="", zlab="")
-        text3d(spec[, 1], spec[, 2], spec[, 3], texts = paste(1:dim(spec)[1]), 
-               adj = 1.3, pos = 4)
-        for(i in which(!is.na(selected))){
-        points3d(spec[i,1],spec[i,2],spec[i,3],
-                 size=10,color=module[which(selected[i]==module[,1]),2], add=T)}
-      }
-    }
+    stop("3D no longer supported.") 
   } 
   return(as.vector(selected))
 }