+++ R/deprecated.defunct.r
@@ -1,5 +1,12 @@
 #' Deprecated and defunct functions in geomorph
 #'
-#' The following function is now defunct
-#'
+#' The following functions are now deprecated
 #' 
+#' build.template
+#' digit.fixed
+#' digitsurface
+#' editTemplate
+#' 
+#'  Starting with geomorph version 4.1, this function is deprecated, as 
+#'  RGL is not supported under MacOS Tahoe. 
+#'  @export